ipa: libipa: agc: Pass the frame duration to configure()
Pass the min/max sensor frame duration to the AGC configure()
operation.

This prepares for limiting the exposure time to the frame duration.

In order to be able to provide to the AGC the frame duration in the IPU3
and Mali-C55 IPAs it is necessary to store the FrameDurationLimits in
the context sensor configuration, like it was already done for the
RkISP1 IPA.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>