ipa: libipa: agc: Rework setLimits() interface
The setLimits() interface accepts a min/max exposure and gains and
detect a request for a fixed value if min == max.

For the shutter time, the concept of max exposure is ill-formed as the
maximum achievable exposure is determinate by the frame duration, while
the min frame duration is a sensor's default parameter.

For gains instead either we want a fixed value or it can range in the
sensor's min and max values.

As we now store the sensor's default in the exposure mode helper, rework
the interface of setLimits() both in AgcMeanLuminance and
ExposureModeHelper so that it receives

- an optional fixed shutter time
- an optional fixed gain
- the current maximum frame duration

If the optional fixed values are populated they are used to fix either
the shutter time or the exposure. If they are not set instead the
algorithm is free to range in the:

- shutterTime = [minExposure, frameDuration - margin]
- gain = [minGain, maxGain]

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>