ipa: libipa: Remove maxExposureTime from sensor configuration
Now that we have moved the AgcMeanLuminance class and the associated
helpers to calculate the maximum achievable shutter time using the
frame duration, remove the maxExposureTime parameter from all the IPAs
using AgcMeanLuminance.

Remove maxExposureTime from all IPA sensor configuration data, as it is
now unused and its value depends on the frame duration.

Do not remove minExposureTime as this instead effectively represents
a sensor parameters which remains stable during a streaming session.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>