ipa: rkisp1: agc: Calculate frame duration using cam helper
The RkISP1 IPA is the only IPA that derives from AgcMeanLuminance
that populates the FrameDuration control.

The way frame duration is calculated is by using the shutter time
as returned by the Agc helpers and use that value to populate the
FrameDuration metadata and calculate the vblank.

This is however not correct as the exposure time shall be shorter than
the frame duration by a sensor-specific margin.

Use the camera sensor helper to calculate the minimum frame duration
required to achieve the newly calculated shutter time.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>