libcamera: request: Remove write-access to controls_
The ControlList inside a Request is initialized with the Camera
ControlInfoMap, which lists the limits of the controls supported
by the Camera.

The Request class currently exposes the control list it contains through
the Request::controls() function as a raw pointer, potentially allowing
applications to replace the Request's control list with a different one.

The following pattern:

	request->controls() = {};

is then currently valid but opens the door to potential issues, as the
new list might be initialized with a different info map or without any
info map at all.

To avoid applications to override the control list contained in a
Request, make the Request::controls() function return a const reference
and move the non-const version to the Request::Private class for
internal usage.

Add to the Request class 4 overloads of the setControl() function that
allows applications to set controls on a Request control list without
accessing it.

Closes: https://gitlab.freedesktop.org/camera/libcamera/-/issues/295
Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>