libcamera: request: Split "Pending" state into two
The "Pending" state at the moment is used for two very distinct scenarios:
when the application alone is managing the request, and when the request
is being processed. This duality makes it harder to track the true state of
a request and consequently it is harder to diagnose misuses.

So split the "Pending" state into the "Idle" and "InProgress" states to
make it explicit when a camera is managing the request. This improves
the detection of request double queueing, but since the status flag is
not updated atomically, it cannot detect all cases.

The name "Pending" is not kept because any application depending on it
has to be adjusted, so it is better to break the API as well. This is also
an ABI break.

Link: https://gitlab.freedesktop.org/camera/libcamera/-/issues/197
Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>