libcamera: device_enumerator_udev: Handle duplicate devices
It is possible that same device is processed multiple times, leading to
multiple `MediaDevice`s being instantiated, mostly likely leading to
a fatal error:

  Trying to register a camera with a duplicated ID xyzw...

This can most easily happen as a result of multiple `enumerate()` calls,
however, currently `enumerate()` is only ever called once, so that won't
trigger the issue.

Nonetheless, it is still possible to trigger this issue due to the fact
that there is a time window after the `udev_monitor` has been created in
`init()` and the first (and only) enumeration done in `enumerate()`. If
e.g. a UVC camera is connected in this time frame, then it is possible
that it will be processed both by the `udevNotify()` and the initial
`enumerate()`, leading to the fatal error. This can be reproduced as
follows:

  1. $ gdb --args cam -m
  2. (gdb) break libcamera::DeviceEnumeratorUdev::enumerate
  3. (gdb) run
  4. when the breakpoint is hit, connect a usb camera
  5. (gdb) continue
  6. observe fatal error

To address this, keep track of the devnums of all devices reported by
udev, and reject devices with already known devnums. This ensures that
any number of `enumerate()` calls and hotplug events will work correctly
(assuming that udev reports "add" / "remove" events correctly).

Closes: https://gitlab.freedesktop.org/camera/libcamera/-/issues/293
Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>