include: linux: drm_fourcc.h: Update to v6.18
Update drm_fourcc.h to Linux kernel version v6.18.

As the upstream drm_fourcc.h version doesn't include formats
definitions for RAW Bayer formats, we have to re-define some of
the symbols we kept downstream as new symbols have been added in
mainline with the same values.

In particular:
-#define IPU3_FORMAT_MOD_PACKED fourcc_mod_code(INTEL, 13)
+#define IPU3_FORMAT_MOD_PACKED fourcc_mod_code(INTEL, 18)

and

-#define DRM_FORMAT_MOD_VENDOR_MIPI 0x0b
-#define DRM_FORMAT_MOD_VENDOR_RPI 0x0c
+#define DRM_FORMAT_MOD_VENDOR_MIPI 0x0e
+#define DRM_FORMAT_MOD_VENDOR_RPI 0x0f

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>