ipa: rkisp1: algorithms: dpf: Use YamlObject::Getter for parsing
Refactor DPF configuration parsing to use YamlObject::Getter
specializations, similar to the approach used in pwl.cpp. This
simplifies parseSingleConfig() by delegating the parsing logic
to dedicated getter specializations.

Implement two YamlObject::Getter specializations:
- rkisp1_cif_isp_dpf_config: Parses filter and nll parameters
- rkisp1_cif_isp_dpf_strength_config: Parses strength parameters

The specializations are defined at the top of the file in the
libcamera namespace.

Remove LOG statements from the getter specializations to avoid
namespace visibility issues, as the LOG macro requires access
to the log category which is defined in a nested namespace.

Signed-off-by: Rui Wang <rui.wang@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>