rkisp1_path: Determine if the ISP is bypassed for YUV and RAW formats
So as to not break existing use cases, in the viewfinder StreamRole RAW
formats that can be converted to NV12 should be preferred, even if the
bits per pixel is lower. Therefore, we should select the RAW format as
our stream format to be configured on the pipeline by default.

However, the user should be given the option to explicitly select if
they wish to use a bypass format if available. This means we have to
enumerate all possible output formats as well as their supported sizes.

Update the rkisp1's generateConfiguration implementation to facilitate
this.

Signed-off-by: Isaac Scott <isaac.scott@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>