rkisp1: Add bypass configurations to ipa_context
The rkisp1 supports ISP bypass for RAW formats. There are other formats
that are not RAW which can also be used in this mode, which are not
currently supported by the rkisp1 pipeline handler.

Add a new IPA context member which tracks formats which are not RAW, so
we can attempt to use them in bypass mode if they are supported.

Signed-off-by: Isaac Scott <isaac.scott@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>