libcamera: pipeline: uvcvideo: Report `FrameDuration`
If available, query the time-per-frame parameter from the device after
starting, and report it in the metadata.

The reported frame duration remains constant during a particular streaming
session, and will most likely not accurately reflect the actual frame
duration for any given frame, depending on the manual exposure time, or if
`V4L2_CID_EXPOSURE_AUTO_PRIORITY` or similar is in effect. But at least it
shows the "intended" frame duration.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>