libcamera: pipeline: uvcvideo: Handle `FrameDurationLimits`
Setting a frame rate is possible with UVC devices using `VIDIOC_S_PARM`.
However, compared to other platforms supported by libcamera, limitations
apply. Probably most crucially, the desired frame rate cannot be set
while the camera is streaming. Furthermore, it is only a single value
and not an allowed range.

So this change only adds support for `FrameDurationLimits` in the control
list passed to `Camera::start()`, and the control is otherwise ignored in
requests.

The kernel interface also only allows a single number and not a range,
so the midpoint of the desired range is used. Checking the supplied
values is not necessary since the kernel will adjust the value if
it is not supported by the device.

Initially the global min/max values are advertised in the `ControlInfo`
of the `FrameDurationLimits` control, which are then updated after
the camera is configured. Updating the control limits after configuration
matches the behaviour of other platforms.

While the kernel interface differentiates three types of frame intervals
(discrete, continuous, stepwise), when querying the available frame intervals
for a given (pixel format, size) combination, all options are evaluated
and only the "local" minimum and maximum is used, as that is the only way
the limits can reasonably be advertised on the libcamera side.

Closes: https://gitlab.freedesktop.org/camera/libcamera/-/issues/296
Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>