libcamera: controls: ControlList::set(): Ensure init list size and type
The initializer list should have the same type as the control, otherwise
the `ControlValue` will have a type different from what is indicated in
the template parameter of the control. This leads to an assertion failure
when the control is later retrieved. Ensure that by using the same template
parameter for the init list.

The same applies to the size of the init list. If the size of the control
is static, then the init list must have the same size. Ensure that by
adding an `assert()` for statically sized controls.

For example, previously

  list.set(controls::FrameDurationLimits, { 1000, 2000 });

would lead to an assertion failure when the control was retrieved since the
deduced type in the init list would be `int`, making the `ControlValue` be of
`ControlTypeInteger32` instead of the expected `ControlTypeInteger64`.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>