ipa: rkisp1: Allow algorithms to update `ControlInfoMap`
In `IPARkISP1::init()`, the `ControlInfoMap` that will be returned
to the pipeline handler is updated after creating the algorithms;
naturally, since otherwise no algorithm could register its own controls.

However, in `IPARkISP1::configure()`, the update is done before the
algorithms are configured. This is not ideal because this prevents
algorithms from correctly updating e.g. the `ControlInfo`s of their
controls.

While no algorithm is affected today, during development, it is useful
for various debug related controls, whose ranges depend on the exact
configuration.

However, `updateControls()` cannot simply be moved to be after the loop
because it updates the limits of `FrameDurationLimits`, which `Agc::configure()`
needs to access. Moving it after the loop would lead to stale data
being used.

So move the updating of `ipaControls` out of `updateControls()`.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>