libcamera: software_isp: Always report saturation metadata
When the saturation is applied the metadata is reported for one frame
only.  Once the colour temperature and saturation is stable, the
metadata for the saturation is not stored into the frame context and
therefore not reported in the completed request metadata.

Update the CCM component to also store the applied saturation in the
frameContext so that it can be reported in the metadata.

Fixes: 59ac34b728c5 ("libcamera: software_isp: Add saturation control")
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>